+++ src/BoclipsClient.ts
@@ -5,6 +5,7 @@ import { HttpFeedsClient } from './sub-clients/httpFeeds/client/HttpFeedsClient'
 import { JobsClient } from './sub-clients/jobs/client/JobsClient';
 import { LegalRestrictionsClient } from './sub-clients/legalRestrictions/client/LegalRestrictionsClient';
 import { OrdersClient } from './sub-clients/orders/client/OrdersClient';
+import { OrganisationAccountsClient } from './sub-clients/organisations/client/OrganisationAccountsClient';
 import { SubjectsClient } from './sub-clients/subjects/client/SubjectsClient';
 
 export interface BoclipsClient {
@@ -16,4 +17,5 @@ export interface BoclipsClient {
   ordersClient: OrdersClient;
   eventsClient: EventsClient;
   jobsClient: JobsClient;
+  organisationAccountsClient: OrganisationAccountsClient;
 }