+++ src/sub-clients/adminLinks/pact/AdminLinksInteractions.ts
@@ -85,6 +85,10 @@ export const getAdminLinks = (): InteractionObject => ({
           href: `${provider.mockService.baseUrl}/v1/events/page-render`,
           templated: false,
         }),
+        independentOrganisations: like({
+          href: `${provider.mockService.baseUrl}/v1/independent-organisations{?countryCode,page,size}`,
+          templated: true,
+        }),
       },
     },
   },