+++ src/sub-clients/organisations/client/OrganisationAccountsClient.ts
@@ -1,0 +1,14 @@
+import Pageable from '../../common/model/Pageable';
+import { OrganisationAccount } from '../model/OrganisationAccount';
+
+export interface IndependentAccountsFilter {
+  countryCode: 'USA';
+  page: number;
+  size: number;
+}
+
+export interface OrganisationAccountsClient {
+  getIndependentAccounts(
+    independentAccountFilter: IndependentAccountsFilter,
+  ): Promise<Pageable<OrganisationAccount> | null>;
+}