+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Independent Organisations Client
+
 ## [2.3.0] - 2019-12-05
 
 - Fix collection contracts